Exception handling

+++ app/Http/Controllers/CategoriesController.php
@@ -24,33 +24,50 @@ class CategoriesController extends Controller
 
     public function show($category_id)
     {
+        if (Categories::find($category_id) === null)
+            return response([
+                'message' => 'Category does not exist'
+            ], 404);
         return Categories::find($category_id);
     }
 
     public function update(Request $request, $category_id)
     {
-        $data = Categories::find($category_id);
+        if (!$data = Categories::find($category_id))
+            return response([
+                'message' => 'Category does not exist'
+            ], 404);
         $data->update($request->all());
         return $data;
     }
 
     public function destroy($category_id)
     {
+        if (Categories::find($category_id) === null)
+            return response([
+                'message' => 'Category does not exist'
+            ], 404);
         return Categories::destroy($category_id);
     }
 
     public function showCategories($post_id)
     {
-        if (!$data =  DB::table('posts')->where('id', $post_id)->first()->category_id) {
-            return response([
-                'message' => 'Invalid post!'
-            ], 404);
-        }
+        try {
+            if (!$data =  DB::table('posts')->where('id', $post_id)->first()->category_id) {
+                return response([
+                    'message' => 'Invalid post!'
+                ], 404);
+            }
 
-        $result = [];
-        $data = json_decode($data);
-        foreach ($data as $value) {
-            array_push($result, Categories::where('id', $value)->first()->title);
+            $result = [];
+            $data = json_decode($data);
+            foreach ($data as $value) {
+                array_push($result, Categories::where('id', $value)->first()->title);
+            }
+        } catch (\Exception $e) {
+            return response([
+                'message' => $e->getMessage()
+            ], 401);
         }
 
         return response($result);
